Added Sprites & Button Interactions
Added in the rest of the sprites and put them into a sprite array (worked around some issues with Unity 6.2), and tied each of them to a state. Toggled the button on and off when it is the values of each state are maxed or not. Added in logic that makes it so the values don't go past the min nor the max values.